EMP-8: change self_delay to work with 1 or 2 bytes

+++ tokens/tokens.h
@@ -241,7 +241,7 @@ void build_masked_tokens_cust(
   struct Balance_l val_cpfp,
   struct Balance_l bal_min_cust,
   struct Balance_l bal_min_merch,
-  uint32_t self_delay,
+  uint16_t self_delay,
 
   struct CommitmentRandomness_l revlock_commitment_randomness_l,
   struct State_l w_new,
@@ -311,7 +311,7 @@ void build_masked_tokens_merch(
   struct Balance_l val_cpfp,
   struct Balance_l bal_min_cust,
   struct Balance_l bal_min_merch,
-  uint32_t self_delay,
+  uint16_t self_delay,
 
   struct HMACKey_l hmac_key,
   struct Mask_l merch_mask_l,
@@ -356,7 +356,7 @@ void issue_tokens(
   Balance_l val_cpfp,
   Balance_l bal_min_cust,
   Balance_l bal_min_merch,
-  uint32_t self_delay,
+  uint16_t self_delay,
   BitcoinPublicKey_l merch_escrow_pub_key_l,
   BitcoinPublicKey_l merch_dispute_key_l,
   BitcoinPublicKey_l merch_payout_pub_key_l,
